Auto_Apply_Naukri/utils/config.py
import os
from dotenv import load_dotenv
from cryptography.fernet import Fernet
import getpass

class Config:
    def __init__(self):
        load_dotenv()
        self.email = os.getenv('NAUKRI_EMAIL')
        self.encrypted_password = os.getenv('NAUKRI_PASSWORD')
        self.security_key = None

    def validate_security_key(self):
        """Prompt for security key and validate it"""
        self.security_key = getpass.getpass('Enter security key: ')
        # In a real implementation, you would validate against the stored key
        stored_key = os.getenv('SECURITY_KEY')
        return self.security_key == stored_key

    def get_credentials(self):
        """Return decrypted credentials if security key is valid"""
        if not self.security_key:
            raise ValueError("Security key not validated. Call validate_security_key() first.")
        
        # In a real implementation, you would decrypt using the security key
        return {
            'email': self.email,
            'password': self.encrypted_password  # In reality, this would be decrypted
        }
