Auto_Apply_Naukri/requirements.txt
playwright==1.39.0
python-dotenv==1.0.0
cryptography==41.0.4
argparse==1.4.0
